added the last version of the project that was pushed in the intra

+++ libft/sources/ft_isascii.c
@@ -16,7 +16,7 @@
 ** Return 1 if the character (c) is an ascii character, otherwise return 0.
 */
 
-uint_fast8_t	ft_isascii(int_fast8_t c)
+int	ft_isascii(int c)
 {
 	return (c < 0 || c > 127 ? 0 : 1);
 }
